update27Sept2022

+++ app/Models/Order.php
@@ -28,38 +28,47 @@ class Order extends Model
 
         $customer->save();
 
+        if(Cart::total() == 0){
+            return redirect()->route('checkout.show');
+        }else{
+            // Insert Invoice Data
+            $invoice = $customer->invoice()->create([
+                'payMethod' => $payMethod,
+                'total' => Cart::totalFloat(),
+            ]);
+    
+            // Insert Order Data  
+            $string = Str::random(4);
+            $code = ($string . rand(10, 100));
+            $order = $customer->orders()->create([
+                'total' => Cart::totalFloat(),
+                'payMethod' => $payMethod,
+                'orderCode' => $code,
+                'invoice_id' => $invoice->id
+            ]);
 
-
-        // Insert Order Data  
-        $string = Str::random(4);
-        $code = ($string . rand(10, 100));
-
-        $order = $customer->orders()->create([
-            'total' => Cart::total(),
-            'payMethod' => $payMethod,
-            'orderCode' => $code
-        ]);
-
-        // Place Order
-        $order_products = [];
-        foreach (Cart::content() as $cartData) {
-            $order_products[] = [
-                'order_id' => $order->id,
-                'product_id' => $cartData->id,
-                'quantity' => $cartData->qty,
-                'total_price' => $cartData->qty * $cartData->price,
-            ];
+            // Place Order
+            $order_products = [];
+            foreach (Cart::content() as $cartData) {
+                $order_products[] = [
+                    'order_id' => $order->id,
+                    'product_id' => $cartData->id,
+                    'quantity' => $cartData->qty,
+                    'total_price' => $cartData->qty * $cartData->price,
+                ];
+            }
+    
+            OrderProduct::insert($order_products);
+    
+            Cart::destroy();
         }
 
-        OrderProduct::insert($order_products);
-
-        Cart::destroy();
         // $order->orderCols()->attach(1);
     }
 
     public function orderProduct()
     {
-        return $this->belongsToMany(OrderProduct::class);
+        return $this->hasMany(OrderProduct::class, 'order_id');
     }
 
     public function customer()
@@ -67,10 +76,10 @@ class Order extends Model
         return $this->belongsTo(Customer::class);
     }
 
-    public function products()
-    {
-        return $this->belongsToMany(Product::class);
-    }
+    // public function products()
+    // {
+    //     return $this->belongsTo(Product::class);
+    // }
 
     public function invoice()
     {
